requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
blinker==1.8.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environ==1.0
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.3
flatbuffers==23.5.26
fsspec==2023.12.2
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
h5py==3.10.0
huggingface-hub==0.23.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
JSON-log-formatter==1.0
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.2
passlib==1.7.4
pillow==10.2.0
protobuf==4.23.4
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-surprise==1.1.3
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
termcolor==2.4.0
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
validators==0.22.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
